requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.1
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
requests==2.28.2
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
six==1.16.0
threadpoolctl==3.1.0
urllib3==1.26.14
